Moved getStrictSchema to parsers. Repurposed utils/types as /async.
