+++ ui/src/adapters/api/schemas.ts
@@ -8,10 +8,10 @@ import {
   buildAPIError,
   buildAuthorizationHeader,
 } from "src/adapters/api";
+import { getStrictParser } from "src/adapters/parsers";
 import { Env, JsonLdType } from "src/domain";
 import { Schema, SchemaAttribute } from "src/domain/schema";
 import { API_VERSION, QUERY_SEARCH_PARAM } from "src/utils/constants";
-import { getStrictParser } from "src/utils/types";
 
 interface Schemas {
   errors: z.ZodError<Schema>[];
