+++ ui/src/adapters/env.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
+import { getStrictParser } from "src/adapters/parsers";
 import { Env } from "src/domain";
-import { getStrictParser } from "src/utils/types";
 
 export interface EnvInput {
   VITE_API_PASSWORD: string;
