+++ ui/src/adapters/parsers/forms.ts
@@ -2,11 +2,11 @@ import dayjs, { isDayjs } from "dayjs";
 import { z } from "zod";
 
 import { CredentialAttribute } from "src/adapters/api/credentials";
+import { getStrictParser } from "src/adapters/parsers";
 import { CredentialForm, CredentialFormAttribute } from "src/domain/credential";
 import { SchemaAttribute } from "src/domain/schema";
 import { ACCESSIBLE_UNTIL } from "src/utils/constants";
 import { formatDate } from "src/utils/forms";
-import { getStrictParser } from "src/utils/types";
 
 // Types
