+++ ui/src/components/credentials/LinkTable.tsx
@@ -19,6 +19,7 @@ import { z } from "zod";
 
 import { APIError } from "src/adapters/api";
 import { OldCredential, credentialUpdate, credentialsGetAll } from "src/adapters/api/credentials";
+import { getStrictParser } from "src/adapters/parsers";
 import { ReactComponent as IconDots } from "src/assets/icons/dots-vertical.svg";
 import { ReactComponent as IconLink } from "src/assets/icons/link-03.svg";
 import { ErrorResult } from "src/components/shared/ErrorResult";
@@ -26,16 +27,11 @@ import { NoResults } from "src/components/shared/NoResults";
 import { TableCard } from "src/components/shared/TableCard";
 import { useEnvContext } from "src/contexts/env";
 import { Schema } from "src/domain/schema";
+import { AsyncTask, isAsyncTaskDataAvailable, isAsyncTaskStarting } from "src/utils/async";
 import { isAbortedError, makeRequestAbortable } from "src/utils/browser";
 import { ACCESSIBLE_UNTIL, LINKS, QUERY_SEARCH_PARAM } from "src/utils/constants";
 import { processZodError } from "src/utils/error";
 import { formatDate } from "src/utils/forms";
-import {
-  AsyncTask,
-  getStrictParser,
-  isAsyncTaskDataAvailable,
-  isAsyncTaskStarting,
-} from "src/utils/types";
 
 const SHOW_SEARCH_PARAM = "show";
